+++ CHANGELOG.md
@@ -3,10 +3,15 @@ Changelog
 
 ## 1.7.0-SNAPSHOT (current master)
 
+### New Features
+
+* add `contributions/count[/density]groupBy/boundary` endpoints to fetch contribution counts/densities grouped by boundary ([#217])
+
 ### Bug Fixes
 
-* fix bug which prevented error messages from being returned for non-unique parameters in POST requests. ([#228])
+* fix bug which prevented error messages from being returned for non-unique parameters in POST requests ([#228])
 
+[#217]: https://github.com/GIScience/ohsome-api/issues/217
 [#228]: https://github.com/GIScience/ohsome-api/pull/228