+++ CHANGELOG.md
@@ -10,9 +10,11 @@ Changelog
 ### Bug Fixes
 
 * fix bug which prevented error messages from being returned for non-unique parameters in POST requests ([#228])
+* fix a regression causing `…/groupBy/boundary` to crash when used with a geometry type filter using the (deprecated) types/keys/values parameter syntax ([#230])
 
 [#217]: https://github.com/GIScience/ohsome-api/issues/217
 [#228]: https://github.com/GIScience/ohsome-api/pull/228
+[#230]: https://github.com/GIScience/ohsome-api/pull/230
 
 
 ## 1.6.0